+++ subgraphs/ethereum/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 5.6.0 (2025-04-16)
 
 - Adds support for Guardian Olympus sUSDS Vault
+- Exclude balances from the Convex Allocator going forward (bricked)
 
 ## 5.5.4 (2025-02-18)